% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{LarvaStage}
\alias{LarvaStage}
\title{Estimate larva stage}
\usage{
LarvaStage(folder_nr, settings)
}
\arguments{
\item{folder_nr}{Folder to test}
}
\description{
This function returns an estimated larva stage for a folder number.
}
