% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Length_median}
\alias{Length_median}
\title{Calculate larval length}
\usage{
Length_median(Data_Table, bin_time)
}
\arguments{
\item{Data_Table}{Data}

\item{bin_time}{Binning window}
}
\value{
A vector of movement
}
\description{
This function calculates movement.
}
