% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{loadAllFiles}
\alias{loadAllFiles}
\title{Load all .csv files}
\usage{
loadAllFiles(larvalcount, inpath, outpath, settings)
}
\arguments{
\item{larvalcount}{How many objects are there}

\item{inpath}{Where to load the data from}

\item{outpath}{Where to save the data to}

\item{settings}{The settings to use for processing}
}
\description{
This function loads .csv files generated by ImageJ/Fiji and processes them
}
