% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Movement_Speed}
\alias{Movement_Speed}
\title{Calculate movement}
\usage{
Movement_Speed(Data_Table, bin_time)
}
\arguments{
\item{Data_Table}{Data}

\item{bin_time}{Binning window}
}
\value{
A vector of movement
}
\description{
This function calculates movement.
}
