% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePlots.R
\name{plotAll}
\alias{plotAll}
\title{Create all plots}
\usage{
plotAll(larvalcount, targetpath, binwindow = 180)
}
\arguments{
\item{larvalcount}{Number of objects to process}

\item{targetpath}{Where to load the data from}

\item{binwindow}{bin width}
}
\description{
Create all plots and save all output
}
