% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePlots.R
\name{plotLength}
\alias{plotLength}
\title{Calculate larval length}
\usage{
plotLength(larvalcount, targetpath, binwindow = 180)
}
\arguments{
\item{larvalcount}{Number of objects to process}

\item{targetpath}{Where to load the data from}

\item{binwindow}{bin width}
}
\description{
Plot larval length
}
